yolo-service/requirements.txt
fastapi
uvicorn
ultralytics
python-multipart
requests
Pillow
